Routing path restructured. Rendered components refactored so that only two are necessary vs. the three components that were originally written
