+++ src/index.js
@@ -7,7 +7,6 @@ import HomePage 				from './pages/HomePage'
 import Personal 				from './pages/Personal'
 import Contact 					from './pages/Contact'
 import { ContactForm } 	from './pages/Contact/components/contactForm'
-import { SubmitSuccess }	from './pages/Contact/components/submitSuccess'
 import PoopMachine 			from './pages/PoopMachine'
 import ErrorPage 				from './pages/ErrorPage'
 import SpinnerOfDoom 		from './pages/SpinnerOfDoom'
@@ -25,10 +24,6 @@ const router = createBrowserRouter([
 			{
 				path: 'contact_form',
 				element: <ContactForm />,
-			},
-			{
-				path: 'submit',
-				element: <SubmitSuccess />,
 			},
 		],	
 	},
