--- src/pages/Contact/style/submitSuccess.module.css
@@ -1,55 +1,0 @@
-/* FILEPATH: /Users/danielbucket/BucketLimited/projects/MustacheMachine/src/pages/Contact/style/submitSuccess.module.css */
-
-.contactFormLink {
-  background-color: #007bff;
-  color: #fff;
-  margin-top: 4rem;
-  padding: 1rem 5rem;
-  border-radius: 1.5rem;
-  font-size: 5rem;
-  transition: background-color 0.3s ease;
-  font-family: Ubuntu_Reg;
-  border: .5rem solid var(--grayd-a-grey);
-}
-
-.contactFormLink:hover {
-  background-color: #0056b3;
-}
-
-.contactFormLink:active {
-  background-color: #003d80;
-}
-
-.formSubmittedContainer {
-  width: 95%;
-  font-size: 3.25rem;
-  font-family: 'Laila_Med';
-}
-
-.formSubmittedHeader {
-  /* display: flex; */
-  /* flex-direction: column; */
-  /* align-items: center; */
-}
-
-.formSubmittedHeader p {
-  color: #007bff;
-  letter-spacing: -0.1rem;
-  padding-bottom: 1rem;
-}
-
-.landingPageRedirect {
-  margin-top: 3rem;
-  font-size: 3rem;
-  font-family: Ubuntu_Reg;
-}
-
-.landingPageRedirect p {
-
-}
-
-.landingPageRedirect span {
-  font-family: KodeMono_Var;
-  font-weight: 700;
-  
-}
